requirements.txt
python-docx==1.1.0
anthropic==0.18.1
openai==1.12.0
